requirements.txt
SQLAlchemy==2.0.40
fastapi==0.115.12
pydantic==2.11.4
nicegui==2.15.0
httpx==0.28.1
paho-mqtt==2.1.0
uvicorn==0.34.2
psycopg2-binary==2.9.10
netifaces2==0.0.22